init.rc: reset apexd.status before starting apexd
Before starting apexd during the post-fs-data trigger, reset the
"apexd.status" property to ensure that apexd considers itself to be
starting up from scratch, rather than being restarted as a lazy service.

This is needed for FDE devices to boot, since FDE devices run the
post-fs-data trigger twice during boot, with services being stopped in
between.  On the second time, the command 'wait_for_prop apexd.status
activated' was hanging forever.  Note, this is true even though FDE
devices don't support updatable APEXes, as apexd is started regardless.

Resetting this property is also needed for userspace reboot, but the
userspace-reboot-requested trigger already had the needed property
reset.  Remove it and instead rely on the reset in post-fs-data.

This is one of a set of changes that is needed to get FDE working again
so that devices that launched with FDE can be upgraded to Android 12.

Bug: 186165644
Test: Tested FDE on Cuttlefish.  Also tested userspace reboot (with FBE)
Change-Id: I4fa57cf15d77b64d1167eaf966347d2a9d6a9b72