init.rc: reset apexd.status after finishing setting up FDE layer
On FDE devices, reset the apexd.status property after finishing setting
up the FDE layer so that apexd correctly considers itself to be starting
from scratch again, as opposed to being restarted as a lazy service.
Without this, the boot will hang forever on the command 'wait_for_prop
apexd.status activated'.

Note that the userspace-reboot-requested trigger also does this, as that
presumably had this same problem too.

This is one of a set of changes that is needed to get FDE working again
so that devices that launched with FDE can be upgraded to Android 12.

Bug: 186165644
Change-Id: I4fa57cf15d77b64d1167eaf966347d2a9d6a9b72