fs_mgr_fstab: Allowlist overlayfs mount flags
Only allow "lowerdir=" and ignore other flags, such as "upperdir".
Ignore "override_creds=" as it is automatically appended by fs_mgr.
Ignore "seclabel" as it is automatically appended by the selinux driver.

"lowerdir=" can be specified in the fs mount flags as well as
fs_mgr_flags, and have the same effect:

  overlay /mnt/product/test1 overlay lowerdir=/1:/2 first_stage_mount
  overlay /mnt/product/test1 overlay ro first_stage_mount,lowerdir=/1:/2

both results in:

  mount(source=overlay,target=/mnt/product/test1,type=overlay,lowerdir=/1:/2,override_creds=off)

Bug: 188862155
Test: Boot to normal with overlayfs mount entries in first stage fstab
Change-Id: I6d6abd44ab32afadec428005f4aece834f9c8905