trusty: apploader: Add --infinite-timeout-retries option
IPC connections to the apploader have a timeout, and
the apploader tool will exit with an error if a connection
times out. This adds a new --infinite-timeout-retries
command line option to the tool that retries to connect
in case of timeout as many times as needed until the
connection succeeds.

Bug: 189911666
Test: m
Change-Id: I4564ab2e5704c105a0d68935a9a3b8bf3a08725b