libutils Threads add pre/post hook.
If client did not set thread create function
with androidSetCreateThreadFunc, these pre/post
hooks run before and after the entry function
on the child thread.

These hooks are only available on non-Windows.

Note that libutils is in the VNDK.
These hooks are used as an alternative to androidSetCreateThreadFunc.
After the corresponding change in AndroidRuntime, even though
there are no other places that uses
- androidSetCreateThreadFunc
- androidCreateRawThreadEtc
These APIs and these functionalities are kept in case it is used in
prebuilts. We can clean up these APIs in follow-up CLs.

Also add getters for these hooks so that, for clients that don't use
libutils Threads (e.g. they use std::thread), they can access and
execute these hooks to attach / detach Java threads as well.

Test: libutils_test
Bug: 190450693
Change-Id: Ib0034d2ea9bc0516c67ab74210e204c89660ce3a