libprocessgroup: Handle permission change failure in uid/pid hierarchy
When a new process is launched it ensure that all files in the uid/pid
hierarchy are accessible by the user/group of that process. While doing
this it is possible that another process in the same uid group is killed
and its uid/pid directory is removed. The loop trying to change permissions
on just removed directory will error out in this case and the process will
fail to migrate into its own uid/pid directory as a result.
Handle this race by ignoring ENOENT error when trying to change access
permissions of the file/directory which has been there at the start of the
operation but does not exist anymore. A warning will be issued in this case
to allow for easy detection of such races.

Bug: 192421915
Signed-off-by: Suren Baghdasaryan <surenb@google.com>
Change-Id: I182298c36f6b0b4580ab59e440bd3aea16f5fbfe