Revert "Re-land removing RECOVERY_AS_BOOT check for init_first_stage"
This reverts commit 4818f74888bfe528d7a2090d18478c95aa88c497.

The previous commit might introduce a build break on legacy
devices that with smaller /boot partition size, because it
outputs two copies of the init binary when
BOARD_USES_RECOVERY_AS_BOOT is true.

    * $OUT/recovery/root/first_stage_ramdisk/init
    * $OUT/recovery/root/system/bin/init

The first one is not needed because there is already a symlink
under $OUT/recovery/root:  init -> /system/bin/init.

Bug: 193565271
Test: make bootimage_debug
Change-Id: I6eca6d8aa66cb0215d7b9dc3f5121dbd704782a2