[fs_mgr] Use shared libutils for remount
remount depends on libbinder.so which already depends on libutils.so.
When linking remount with libutils.a, the linker tries to open objects
from libutils.a to define undefined symbols in libbinder.so.  This
causes linker failure with upstream LLD.  Switching to shared
libutils.so instead doesn't increase the dependency closure for remount.

Test: build with aosp/1809741 which has lld with the above behavior.
Change-Id: I858042a08babe2c2db1fc8c7eeb5de62dda5f328