libsnapshot: Verify merge sequence after parsing
This adds verification to ensure the merge sequence does not attempt to
use blocks that are already overwritten. It will print the first
conflict in the logs. Additionally, adds a forwards merge iterator to
cow reader, as well as an option to use it in inspect_cow.

This can be accessed via the VerifyMergeOps function in Cow Reader, or
by using inspect_cow -v [COWFILE]

Bug: 200076590
Test: cow_api_test.InvalidMergeOrderTets
Change-Id: I893b9f8a8803cb6dd53225ec34224167b9fe2fda