libsnapshot: Add Forward Merge Iterator with debug
This adds a forwards merge iterator, and adds it to inspect_cow.
Additionally, allow inspect_cow to verify that the merge sequence is
valid (does not overwrite needed sources). This requires allowing
iterators to look at already merged items, so pass a debug flag to the
iterators, and don't delete the already merged ops

Bug: 200076590
Test: inspect_cow -v -n [cow_file], cow_api_test
Change-Id: I893b9f8a8803cb6dd53225ec34224167b9fe2fda