storageproxyd: Request sense data on UNIT ATTENTION
Adds a REQUEST SENSE command after an RPMB command triggered a UNIT
ATTENTION condition. Adds retries for the RPMB response request and
read after a UNIT ATTENTION has been addressed. The initial write will
not be retried as we are not sure if the RPMB write has been committed
after a UNIT ATTENTION response.

Test: m storageproxyd
Bug: 200037389
Change-Id: Ib970e779365bb396756c114684c7f6e56a737ab1