storageproxyd: stall on RPMB writes when checkpointing, if necessary
If a checkpointing operation is in progress, RPMB_SEND operations will
be stalled for any write that is flagged as
STORAGE_MSG_FLAG_PRE_COMMIT_CHECKPOINT. This avoids the awkward case
where the superblock is committed, but then the data is not because
the checkpoint is rolled back.

Based on Stephen's CL/1845477 "Add helper to check checkpoint state of
mounts".

Test: m storageproxyd
Bug: 194313068
Change-Id: Ib6a432db1bc1b034f803b743b0d7322e3f31d814