stub_zygote: add.
The 32-bit zygote costs a significant amount of memory, even in the
case where there are zero 32-bit applications running. As an
intermediate step to completely killing 32-bit, we can win back
the memory usage in the common case by spawning the 32-bit zygote on
demand. Implement a server that listens on the zygote socket and spawns
a new zygote for each request to spawn a 32-bit application.

Test: manual via editing rootdir/init.rc to import the zygote64_stub32.rc,
      since ro.zygote is set in the vendor image
Change-Id: I905692f18568caa6017356bfc16f10589abc0a28