canned fs_config accepts glob pattern
Previously, each file/dir in a filesystem had to have a matching entry
in the canned fs_config file. With this change, the config file can have
glob patterns. This not only makes it easier to writhe the config file,
but also allows us to customize a generic fs_config file by appending
more specific patterns to the file. For example:

/lib/* 1000 1000 0644  // a generic fs_config entry
/lib/libfoo.so 1000 2000 0600 // customized fs_config entry

Bug: 209971551
Test: m
Change-Id: I43902fed08db1b4968d02c75fac0a47976fff72a