Let snapuserd started before init second stage
1. snapuserd need to be put to recovery/root/first_stage_ramdisk/system/bin
when BOARD_USES_RECOVERY_AS_BOOT := true is set, or it will appear the
following errors:
[   11.602223] init: Relaunched snapuserd with pid: 152
[   11.602700] init: Creating logical partitions with snapshots as needed
[   11.603555] init: Cannot launch snapuserd; execv failed: No such file or directory

2. snapuserd need to be started before the init second stage, or it will
appear the following errors. It because if init second stage is started before snapuserd,
/dev/__properties__ will be created by init. Then in the function of
__system_properties_init during the libc init stage of snapuserd process, it will
go to ContextsSplit.Initialize finally, and in this function, it will try to read
/system/etc/selinux/plat_property_contexts. However, system partition should be
prepared by snapuserd which is not started yes, this will halt both init and snapuserd
processes:
[   12.367545] init: Relaunched snapuserd with pid: 211
[   12.382032] init: init second stage started!
[   16.434536] I/O error: sector 550576: no user-space daemon for dm-user/system_b-user-cow target pid:128 tgid:128
[   16.436606] I/O error: sector 550568: no user-space daemon for dm-user/system_b-user-cow target pid:128 tgid:128
[   16.439396] I/O error: sector 550560: no user-space daemon for dm-user/system_b-user-cow target pid:128 tgid:128
[   16.441467] I/O error: sector 550552: no user-space daemon for dm-user/system_b-user-cow target pid:128 tgid:128
[   20.466540] I/O error: sector 5792: no user-space daemon for dm-user/system_b-user-cow target pid:128 tgid:128
[   20.468463] device-mapper: verity-fec: 252:4: FEC 2949120: read failed (720): -5 pid:125 tgid:125

It because if init second stage is started before snapuserd, /dev/__properties__
will be created by init. Then in the function of __system_properties_init during
the libc init stage of snapuserd process, it will go to ContextsSplit.Initialize
finally, and in this function, it will try to read /system/etc/selinux/plat_property_contexts.
However, system partition should be prepared by snapuserd which is not started yes,
this will halt both init and snapuserd processes

Bugnizar ID: 216026310

Test: manual

Signed-off-by: ji, zhenlong z <zhenlong.z.ji@intel.com>
Change-Id: Ib736da1a43db4996a20e680b8be1f99dfe13ec9d