Move the blkio properties into task_profiles.json
This patch prepares for the migration of the blkio controller to the v2
cgroup hierarchy. The foreground BFQ weight in task_profiles.json is the
default BFQ weight. From
https://www.kernel.org/doc/Documentation/block/bfq-iosched.txt:

weight (namely blkio.bfq.weight or io.bfq-weight): the weight of the
group inside its parent. Available values: 1..10000 (default 100). The
linear mapping between ioprio and weights, described at the beginning
of the tunable section, is still valid, but all weights higher than
IOPRIO_BE_NR*10 are mapped to ioprio 0.

Bug: 213617178
Test: Booted Android in Cuttlefish and verified the blkio cgroup settings.
Change-Id: If8bfa67d74b3457bc8d512fde70cd5e01d4c60fd
Signed-off-by: Bart Van Assche <bvanassche@google.com>