Add ATrace events to libtrusty
Add ATrace tracking spans to `tipc_connect` and `tipc_send` in libtrusty
in order to provide better tracing around tipc communcation in userspace
apps. This adds libcutils as a dependency for libtrusty, which in turn
requires that it be added as a dependency for the storage proxy and
trusty-uti-ctrl pacakges since they depend on libtrusty.

Test: Manual - verified that trace works with emulator
Bug: 211044277
Change-Id: I374ff7489f1cd13a7d35700a03ff1bac738cfa4c