Add support for tombstone symbolization to pbtombstone.
This patch teaches pbtombstone to use llvm-symbolizer to symbolize
stack traces and augment the protobuf tombstones with the symbol
information, before printing tombstones with the symbolized stack
traces included.

The main advantage of adding this information to the tombstone
as opposed to having developers use the stack tool is that stack
does not print all of the information in the original tombstone,
which means that both reports may be required to understand a crash.
Furthermore, stack traces printed by stack are not correlated with
the stack traces in the tombstone, making the report harder to read,
especially with GWP-ASan and MTE which may produce multiple stack
traces for the crashing thread.

Although we could teach stack to print more information, this would
continue to be fragile because stack relies on parsing textual
tombstones. Switching stack to read proto tombstones would be
tantamount to a full rewrite and would require duplicating the C++
proto-to-text logic that we already have in Python. It seems better
to reuse the C++ code for the proto-based symbolization tool.

Having the ability to store symbol information in tombstones
also opens up some possibilities for sharing symbol information
with developers. For example, this makes it possible to provide
partially symbolized (using system binaries) tombstones to application
developers, who would in turn be able to symbolize frames from their
application binaries.

The symbols may come from two possible sources:

- The local symbols tree, when running in a lunched environment.

- A symbol server. Recent versions of llvm-symbolizer (although not the
  version currently released in AOSP) have the ability to query
  a symbol server when given a build id. This can either be a
  local symbol server (e.g. pointed to an app's build directory)
  or a remote symbol server (e.g. hosting CI builds). The elfutils
  project has a compatible symbol server named debuginfod. There is
  also an llvm-debuginfod in development.

This is a rough draft of what I expect this to look like. The
user interface may need some tweaking. For example, we may want
symbolization to be done in a separate tool.

Change-Id: Ia4676821cf980c69487cf11aefa2a02dc0c1626f