chmod ug+x to /dev/cpuctl
So far, the directory didn't have the execute bit on. Therefore only the
root could have access to files and dirs under the directory. Now, at
least crosvm has to access the files and dirs so that I can join itself
to a specific cgroup according to the selected task profile.

Bug: 223790172
Test: adb shell /apex/com.android.compos/bin/composd_cmd test-compile
Change-Id: Ie2cdebc1ae5890a1bd9e9f2a7cd9273c62bf022a