libprocessgroup: ignore failure to adjust permissions if file is gone
Occasional boot failures have been reported which are caused by a failure
to adjust file permissions in a newly created cgroup. Current theory is
that the cgroup is being concurrently deleted. Add checks to confirm or
refute this theory. Once the reason is clarified the checks should be
replaced by a proper fix.

Bug: 228160715
Signed-off-by: Suren Baghdasaryan <surenb@google.com>
Change-Id: I88e03c73c61b70276bde405f9abdd38903db3641