WIP: libprocessgroup: Handle temporarily non-empty group removals
Cgroup removal fails with EBUSY if there are active processes or threads
left alive. Occasionally a thread or a process might be stuck in an
interruptible sleep and take some time during exit. In such cases attempts
to remove the cgroup it belongs to will fail.
RemoveProcessGroup implements a retry mechanism which tries to remove
the cgroup for 200ms and then gives up. This results in occasional
leftover cgroups. These empty unused cgroups consume memory.
Clean up these leftover cgroups by storing the ones we failed to remove
and removing the next time RemoveProcessGroup gets called.

Bug: 233319780
Signed-off-by: Suren Baghdasaryan <surenb@google.com>
Change-Id: I2e4bb1f7b7e19c904c85faea7bbabbfdef9c8125