libsnapshot: Disable userspace snapshot for GRFd devices with Vendor on S
Earlier we used to check only
ro.vendor.build.version.release_or_codename if vendor is Android 12;
however that is no sufficient.

Disable userspace snapshots if ro.vndk.version is set to 32.

If the vendor partition is on S and system partition is on T,
certain tests in vts_libsnapshot_test used to fail. This is primarily
because of inconsistent check between daemon and vts test.

Bug: 231401995
Test: vts_libsnapshot_test on S vendor and T system
      vts_libsnapshot_test on T vendor and T system
Signed-off-by: Akilesh Kailash <akailash@google.com>
Change-Id: Iad4f299bd2e07c9c01f5fbee6a20e2f01bf1778a