Light Weight Threads (LWT) initial version.
This is a second attempt at putting LWT into Android,
temporarily not in its final location.  An earlier
attempt at doing this was done with about 45 commits,
but the "repo upload . --no-verify" (an attempt to
bypass the clang-format issues until later) resulted
when trying to just "repo upload ." after addressing
the clang-format issues in an initial way would fail
because the script would run for each one of the 45
commits.  To work around that the commits where moved
to the earlier location of this work in Git-on-Borg at:

    https://user.git.corp.google.com/pantin/lwt/

Test: ./mk x64
Test: ./run x64
Test: ./mk arm64
Test: ./run arm64

Signed-off-by: Ramon Pantin <pantin@google.com>
Change-Id: I3c1e74cb8ee4b59bed6d70c3d13a49a9117e4664