Refactor ScudoCrashData class into a function.
The original refactor was reverted, so this redoes the refactor.
The previous problem where the fault address is unreadable is fixed.

The ScudoCrashData class isn't strictly necessary, so remove it.
Change the memory data to a vector that is initialized to zero,
since the always zero scudo option might be turned off.
Modify algorithm to try and read the data around the fault
address in a large chunks, instead of reading a page at a time.

Test: All unit tests pass (32 bit and 64 bit).
Change-Id: I2ac7bbe11d4e592cc097c3bc527c10a1c5bb50c1